requirements.txt
flask==2.2.2
openai==0.25.0
gunicorn==20.1.0
google-cloud-logging==3.3.1